dsm: refactor so that unit test can run against real code
Starting to get to a pattern that allows flight unit tests... and
approaching a time for a lot more makefile refactoring (next rel).
