+++ flight/targets/flightd/fw/pios_config.h
@@ -108,6 +108,7 @@
 #define PIOS_INCLUDE_HMC5983
 #endif
 
+#define PIOS_INCLUDE_FAKETICK
 #define PIOS_INCLUDE_SIMSENSORS
 #define PIOS_INCLUDE_ADC
